+++ libs/chat/src/lib/primitives/chat-message-list/message-template.directive.ts
