+++ libs/chat/src/lib/styles/chat-message-list.styles.ts
@@ -1,0 +1,14 @@
+// libs/chat/src/lib/styles/chat-message-list.styles.ts
+// SPDX-License-Identifier: MIT
+export const CHAT_MESSAGE_LIST_STYLES = `
+  :host {
+    display: flex;
+    flex-direction: column;
+    gap: 0;
+    padding: var(--ngaf-chat-space-4) var(--ngaf-chat-space-6);
+    max-width: var(--ngaf-chat-max-width);
+    margin: 0 auto;
+    width: 100%;
+    box-sizing: border-box;
+  }
+`;
